+++ src/api/pbx.ts
@@ -392,6 +392,7 @@ export class PBX extends EventEmitter {
         'p4_ptype',
         'pnumber',
         'language',
+        'phoneappli.enable',
       ],
     })
 
@@ -424,6 +425,7 @@ export class PBX extends EventEmitter {
       name: userName,
       phones,
       language: lang,
+      phoneappli: toBoolean(res?.[7]),
     }
   }
 
@@ -463,6 +465,24 @@ export class PBX extends EventEmitter {
     }
     const res = await this.client.call_pal('getPhonebooks')
     return res?.filter(item => !item.shared) || []
+  }
+  getPhoneappliContact = async (tenant: string, user: string, tel: string) => {
+    if (this.isMainInstance) {
+      await waitPbx()
+    }
+    if (!this.client) {
+      return
+    }
+    try {
+      const res = await this.client.call_pal('getPhoneAppliContact', {
+        tenant,
+        user,
+        tel,
+      })
+      return res
+    } catch (err) {
+      return {}
+    }
   }
   getContact = async (id: string) => {
     if (this.isMainInstance) {