+++ src/brekekejs/index.d.ts
@@ -171,6 +171,11 @@ export type PbxPal = {
     resolve: (res: PbxContact) => void,
     reject: ErrorHandler,
   ): void
+  getPhoneAppliContact(
+    p: PbxGetPhoneappliContactParam,
+    resolve: (res: PbxPhoneappliContact) => void,
+    reject: ErrorHandler,
+  ): void
   setContact(
     p: PbxContact,
     resolve: (res: PbxContact) => void,
@@ -273,6 +278,11 @@ export type PbxGetContactListItem = {
   user?: string
 }
 
+export type PbxGetPhoneappliContactParam = {
+  tenant: string
+  user: string
+  tel: string
+}
 export type PbxGetContactParam = {
   aid: string
 }
@@ -315,6 +325,10 @@ export type PbxParkParam = {
 export type PbxBook = {
   phonebook: string
   shared?: string
+}
+export type PbxPhoneappliContact = {
+  display_name?: string
+  image_url?: string
 }
 export type PbxContact = {
   aid: string