+++ src/components/navigationConfig2.ts
@@ -1,10 +1,12 @@
 import { action } from 'mobx'
 import { ReactComponentLike } from 'prop-types'
 
+import { getAuthStore } from '../stores/authStore'
 import { intl } from '../stores/intl'
 import { RnAlert } from '../stores/RnAlert'
 import { RnStacker } from '../stores/RnStacker'
 import { arrToMap } from '../utils/arrToMap'
+import { openLinkSafely, URLSchemes } from '../utils/deeplink'
 import { Menu, SubMenu } from './navigationConfig'
 
 let PageCallTransferChooseUser: ReactComponentLike
@@ -17,22 +19,32 @@ export const setPageCallTransferDial = (p: ReactComponentLike) => {
 }
 
 export const getTabs = (tab: string) => {
+  const subMenus = [
+    {
+      key: 'list_user',
+      label: intl`USER`,
+      navFnKey: { PageCallTransferChooseUser },
+    },
+    {
+      key: 'external_number',
+      label: intl`KEYPAD`,
+      navFnKey: { PageCallTransferDial },
+    },
+  ]
+  // add phonebook tab if phoneappli is enabled
+  if (getAuthStore().phoneappliEnabled()) {
+    subMenus.push({
+      key: 'phonebook',
+      label: intl`PHONEBOOK`,
+      navFnKey: {} as any,
+    })
+  }
+
   const arr = [
     {
       key: 'call_transfer',
       icon: null,
-      subMenus: [
-        {
-          key: 'list_user',
-          label: intl`USER`,
-          navFnKey: { PageCallTransferChooseUser },
-        },
-        {
-          key: 'external_number',
-          label: intl`KEYPAD`,
-          navFnKey: { PageCallTransferDial },
-        },
-      ],
+      subMenus,
       defaultSubMenuKey: 'list_user',
     },
   ] as any as Menu[]
@@ -49,6 +61,11 @@ export const getTabs = (tab: string) => {
     m.subMenus.forEach(s => {
       s.navFn = action(() => {
         const name = Object.keys(s.navFnKey)[0]
+        // handle link to phoneappli app
+        if (!name || s.key === 'phonebook') {
+          openLinkSafely(URLSchemes.phoneappli.USERS)
+          return
+        }
         // @ts-ignore
         const Component: ReactComponentLike = s.navFnKey[name]
         const lastStack = RnStacker.stacks.pop()